pyproject.toml
[project]
name = "medium-article-writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.13",
    "arxiv>=2.3.1",
    "ddgs>=9.9.3",
    "googlesearch-python>=1.3.0",
    "lxml-html-clean>=0.4.3",
    "newspaper4k>=0.9.4.1",
    "openai>=2.13.0",
    "pycountry>=24.6.1",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "tweepy>=4.16.0",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.3",
]
